+++ test/index.js
@@ -1,24 +1,65 @@
 "use strict";
 
-const assertTransform = require("assert-transform");
-const babel = require("babel-core");
 const path = require("path");
 const assert = require("assert");
+const fs = require("fs");
+const { EOL } = require("os");
+const { promisify } = require("util");
+
+const babel = require("@babel/core");
+const jsdiff = require("diff");
+const chalk = require("chalk");
 
 const babelPluginTransformDefine = require("../lib/index.js");
 
-const getBabelOps = (pluginOps) => {
-  return {
-    "presets": ["es2015"],
-    "plugins": [
-      [path.resolve(__dirname, "../lib/index.js"), pluginOps]
-    ]
+const readFile = promisify(fs.readFile);
+const splitLines = ({ value }, fn) => value
+  .split(EOL)
+  .map((line, idx, lines) => line === "" && idx === lines.length - 1 ? "" : fn(line))
+  .join(EOL);
+
+const assertTransform = async (initial, expected, opts) => {
+  const transformOpts = {
+    ...opts,
+    // Specify filename to pick up local `.babelrc.js`
+    // https://babeljs.io/docs/en/options#babelrc
+    filename: initial
   };
+
+  // Note: We trim + EOL to normalize whitespace + give readable error diffs
+  const [actualCode, expectedCode] = await Promise.all([
+    readFile(initial).then((code) => babel.transform(code, transformOpts).code.trim() + EOL),
+    readFile(expected).then((buf) => buf.toString().trim() + EOL)
+  ]);
+
+  const diff = jsdiff.diffLines(actualCode, expectedCode);
+  // Consider no diff or newline-only diff to be "the same".
+  const changes = diff.filter(({ added, removed, value }) => (added || removed) && value.trim());
+  if (changes.length === 0) {
+    return true;
+  }
+
+  const msg = diff
+    .map((obj) => {
+      if (obj.added) { return splitLines(obj, (line) => chalk `{green +${line}}`); }
+      if (obj.removed) { return splitLines(obj, (line) => chalk `{red -${line}}`); }
+      return splitLines(obj, (line) => chalk `{grey  ${line}}`);
+    })
+    .join("");
+
+  throw new Error(chalk `{white Difference found ({green actual}, {red expected}): ${EOL}}${msg}`);
 };
+
+const getBabelOps = (pluginOps) => ({
+  presets: ["@babel/preset-env"],
+  plugins: [
+    [path.resolve(__dirname, "../lib/index.js"), pluginOps]
+  ]
+});
 
 describe("babel-plugin-transform-define", () => {
   before(function () {
-    // TODO: WTF babel needs to warm up! This is Bullshit!
+    // TODO: See if needed. Previously here to warm up babel for tests.
     this.timeout(10000); // eslint-disable-line
     babel.transform("const x = 1;", getBabelOps());
   });
@@ -37,7 +78,7 @@ describe("babel-plugin-transform-define", () => {
 
       it("should transform with config defined by an Object", () => {
         const babelOpts = getBabelOps({
-          "process": {
+          process: {
             env: {
               NODE_ENV: "development"
             }
@@ -62,8 +103,8 @@ describe("babel-plugin-transform-define", () => {
 
     it("should transform Identifiers", () => {
       const babelOpts = getBabelOps({
-        "VERSION": "1.0.0",
-        "PRODUCTION": true
+        VERSION: "1.0.0",
+        PRODUCTION: true
       });
 
       return assertTransform(
@@ -73,61 +114,58 @@ describe("babel-plugin-transform-define", () => {
 
     it("should transform false", () => {
       const babelOpts = getBabelOps({
-        "PRODUCTION": false
+        PRODUCTION: false
       });
 
       return assertTransform(
         path.join(__dirname, "./false/actual.js"),
         path.join(__dirname, "./false/expected.js"), babelOpts);
-    })
+    });
 
     it("should transform 0", () => {
       const babelOpts = getBabelOps({
-        "PRODUCTION": 0
+        PRODUCTION: 0
       });
 
       return assertTransform(
         path.join(__dirname, "./0/actual.js"),
         path.join(__dirname, "./0/expected.js"), babelOpts);
-    })
+    });
 
     it("should transform empty string", () => {
       const babelOpts = getBabelOps({
-        "PRODUCTION": ''
+        PRODUCTION: ""
       });
 
       return assertTransform(
         path.join(__dirname, "./emptyString/actual.js"),
         path.join(__dirname, "./emptyString/expected.js"), babelOpts);
-    })
+    });
 
     it("should transform null", () => {
       const babelOpts = getBabelOps({
-        "PRODUCTION": null
+        PRODUCTION: null
       });
 
       return assertTransform(
         path.join(__dirname, "./null/actual.js"),
         path.join(__dirname, "./null/expected.js"), babelOpts);
-    })
+    });
 
     it("should transform undefined", () => {
       const babelOpts = getBabelOps({
-        "PRODUCTION": undefined
+        PRODUCTION: undefined
       });
 
       return assertTransform(
         path.join(__dirname, "./undefined/actual.js"),
         path.join(__dirname, "./undefined/expected.js"), babelOpts);
-    })
-
-    it("should transform code from config in a file", () => {
-      const babelOpts = getBabelOps("./test/load-config-file/config.js");
-
-      return assertTransform(
-        path.join(__dirname, "./load-config-file/actual.js"),
-        path.join(__dirname, "./load-config-file/expected.js"), babelOpts);
     });
+
+    it("should transform code from dynamic .babelrc.js", () => assertTransform(
+      path.join(__dirname, "./load-dynamic-babelrc/actual.js"),
+      path.join(__dirname, "./load-dynamic-babelrc/expected.js")
+    ));
   });
 
   describe("unit tests", () => {
@@ -144,11 +182,13 @@ describe("babel-plugin-transform-define", () => {
         objectPaths = babelPluginTransformDefine.getSortedObjectPaths({ process: "env" });
         assert(Array.isArray(objectPaths));
       });
+
       it("should return a complete list of paths", () => {
         const obj = { process: { env: { NODE_ENV: "development" } } };
         const objectPaths = babelPluginTransformDefine.getSortedObjectPaths(obj);
-        assert.deepEqual(objectPaths, [ "process.env.NODE_ENV", "process.env", "process" ]);
+        assert.deepEqual(objectPaths, ["process.env.NODE_ENV", "process.env", "process"]);
       });
+
       it("should return a list sorted by length", () => {
         const obj = { process: { env: { NODE_ENV: "development" } } };
         const objectPaths = babelPluginTransformDefine.getSortedObjectPaths(obj);